# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_jhoicas_Inventario-api_internal_application_dto.CreateUserRequest:
    properties:
      company_id:
        type: string
      email:
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
      password:
        minLength: 8
        type: string
      roles:
        description: roles asignados al usuario
        items:
          type: string
        type: array
    required:
    - company_id
    - email
    - name
    - password
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.CreateWarehouseRequest:
    properties:
      address:
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.UserListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.UserResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_jhoicas_Inventario-api_internal_application_dto.UserResponse:
    properties:
      company_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.WarehouseListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.WarehouseResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_jhoicas_Inventario-api_internal_application_dto.WarehouseResponse:
    properties:
      address:
        type: string
      company_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
host: api.naturerp.ludoia.com
info:
  contact: {}
  description: API para el sistema ERP.
  title: Tu API ERP
  version: "1.0"
paths:
...
  /api/users:
    get:
      description: Lista los usuarios de la empresa autenticada (solo admin)
      parameters:
      - default: 20
        description: Límite
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.UserListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Listar usuarios
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Crea un usuario dentro de la empresa autenticada (solo admin)
      parameters:
      - description: Datos del usuario (password plano, se hashea)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Crear usuario
      tags:
      - users
  /api/users/{id}:
    put:
      consumes:
      - application/json
      description: Actualiza roles o estado de un usuario (solo admin)
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      - description: '{\'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Actualizar usuario
      tags:
      - users
...
  /api/warehouses:
    get:
      parameters:
      - default: 20
        description: Límite
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.WarehouseListResponse'
      security:
      - Bearer: []
      summary: Listar bodegas
      tags:
      - warehouses
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos de la bodega
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.CreateWarehouseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.WarehouseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Crear bodega
      tags:
      - warehouses
  /api/warehouses/{id}:
    get:
      parameters:
      - description: ID de la bodega
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.WarehouseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Obtener bodega por ID
      tags:
      - warehouses
securityDefinitions:
  Bearer:
    description: |-
      Escribe "Bearer " seguido de un espacio y tu token.
      Escribe "Bearer " seguido de un espacio y tu token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"